Complete refactoring pull request
added common image raster, for DRY
fix OCP violation
Fix ADP violation
