+++ jme3-core/src/main/resources/com/jme3/system/version.properties
@@ -1,11 +1,12 @@
 # THIS IS AN AUTO-GENERATED FILE..
 # DO NOT MODIFY!
-build.date=1900-01-01
-git.revision=0
-git.branch=unknown
-git.hash=
-git.hash.short=
-git.tag=
-name.full=jMonkeyEngine 3.1.0-UNKNOWN
+build.date=2016-04-10
+git.revision=5484
+git.branch=master
+git.hash=8a8dd72dc9a5bc1ebdc2ec7d1b174ec5256cb734
+git.hash.short=8a8dd72
+git.tag=null
+name.full=jMonkeyEngine 3.1-5484
+version.full=3.1-5484
 version.number=3.1.0
 version.tag=SNAPSHOT
